requirement.txt
pyserial
requests
thread6
paho-mqtt
google-cloud-texttospeech
python-vlc